Add unit tests

+++ rtpi-api/src/main/kotlin/ie/dublin/rtpi/api/LiveData.kt
@@ -24,7 +24,7 @@ interface TimedLiveData : LiveData {
 }
 
 data class DueTime(
-    val minutes: Long,
+    val minutes: Int,
     val time: LocalTime
 )
